fix must_use error, move shifts to permutation, reorganise code
