+++ kimchi/src/circuits/polynomials/permutation.rs
@@ -51,6 +51,8 @@ use ark_poly::{
     EvaluationDomain, Evaluations, Radix2EvaluationDomain as D,
 };
 use ark_poly::{Polynomial, UVPolynomial};
+use array_init::array_init;
+use blake2::{Blake2b512, Digest};
 use o1_utils::{ExtendedDensePolynomial, ExtendedEvaluations};
 use rand::{CryptoRng, RngCore};
 
@@ -110,6 +112,78 @@ pub fn zk_polynomial<F: FftField>(domain: D<F>) -> DensePolynomial<F> {
         -w1 - w2 - w3,                // x^2
         F::one(),                     // x^3
     ])
+}
+
+/// Shifts represent the shifts required in the permutation argument of PLONK.
+/// It also caches the shifted powers of omega for optimization purposes.
+pub struct Shifts<F> {
+    /// The coefficients `k` (in the Plonk paper) that create a coset when multiplied with the generator of our domain.
+    pub(crate) shifts: [F; PERMUTS],
+    /// A matrix that maps all cells coordinates `{col, row}` to their shifted field element.
+    /// For example the cell `{col:2, row:1}` will map to `omega * k2`,
+    /// which lives in `map[2][1]`
+    pub(crate) map: [Vec<F>; PERMUTS],
+}
+
+impl<F> Shifts<F>
+where
+    F: FftField + SquareRootField,
+{
+    /// Generates the shifts for a given domain
+    pub fn new(domain: &D<F>) -> Self {
+        let mut shifts = [F::zero(); PERMUTS];
+
+        // first shift is the identity
+        shifts[0] = F::one();
+
+        // sample the other shifts
+        let mut i: u32 = 7;
+        for idx in 1..(PERMUTS) {
+            let mut shift = Self::sample(domain, &mut i);
+            // they have to be distincts
+            while shifts.contains(&shift) {
+                shift = Self::sample(domain, &mut i);
+            }
+            shifts[idx] = shift;
+        }
+
+        // create a map of cells to their shifted value
+        let map: [Vec<F>; PERMUTS] =
+            array_init(|i| domain.elements().map(|elm| shifts[i] * elm).collect());
+
+        //
+        Self { shifts, map }
+    }
+
+    /// retrieve the shifts
+    pub fn shifts(&self) -> &[F; PERMUTS] {
+        &self.shifts
+    }
+
+    /// sample coordinate shifts deterministically
+    fn sample(domain: &D<F>, input: &mut u32) -> F {
+        let mut h = Blake2b512::new();
+
+        *input += 1;
+        h.update(&input.to_be_bytes());
+
+        let mut shift = F::from_random_bytes(&h.finalize()[..31])
+            .expect("our field elements fit in more than 31 bytes");
+
+        while !shift.legendre().is_qnr() || domain.evaluate_vanishing_polynomial(shift).is_zero() {
+            let mut h = Blake2b512::new();
+            *input += 1;
+            h.update(&input.to_be_bytes());
+            shift = F::from_random_bytes(&h.finalize()[..31])
+                .expect("our field elements fit in more than 31 bytes");
+        }
+        shift
+    }
+
+    /// Returns the field element that represents a position
+    pub(crate) fn cell_to_field(&self, &Wire { row, col }: &Wire) -> F {
+        self.map[col][row]
+    }
 }
 
 impl<F: FftField + SquareRootField> ConstraintSystem<F> {
